Added permission chceck to UserApi

+++ Source_code/RestAPI/src/src/main/java/io/swagger/api/UserApiController.java
@@ -5,9 +5,10 @@ import io.swagger.model.Borrowed;
 import io.swagger.model.User;
 import io.swagger.model.UserStatus;
 import io.swagger.model.UserToken;
+import io.swagger.model.UserType;
 import io.swagger.service.LibraryService;
 import io.swagger.service.UserService;
-import io.swagger.service.UserTokenService;
+import io.swagger.service.UserTypeService;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.*;
@@ -41,8 +42,6 @@ public class UserApiController implements UserApi {
     @Autowired
     private LibraryService libraryService;
 
-    @Autowired
-    private UserTokenService userTokenService;
     private static final Logger log = LoggerFactory.getLogger(UserApiController.class);
 
     private final ObjectMapper objectMapper;
@@ -58,10 +57,15 @@ public class UserApiController implements UserApi {
     public ResponseEntity<Void> createUser(@ApiParam(value = "Created user object" ,required=true )  @Valid @RequestBody User body,@ApiParam(value = "") @Valid @RequestParam(value = "token", required = false) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            if(userService.createUser(body) != null){
-                return new ResponseEntity<Void>(HttpStatus.OK);
+            if(userService.isModifyPermittedForToken(body, token)){
+                body = userService.initNewUserAccordingToPermissions(body, token);
+                if(userService.createUser(body) != null){
+                    return new ResponseEntity<Void>(HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+                }
             } else {
-                return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+                return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
             }
         } else {
             return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
@@ -72,11 +76,15 @@ public class UserApiController implements UserApi {
     public ResponseEntity<Void> deleteUserById(@ApiParam(value = "The Id that needs to be deleted",required=true) @PathVariable("id") Long id,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "token", required = true) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            User user = userService.deleteUserById(id);
-            if (user != null) {
-                return new ResponseEntity<Void>(HttpStatus.OK);
-            } else { 
-                return ResponseEntity.notFound().build();
+            if(userService.isModifyPermittedForToken(id, token)){
+                User user = userService.deleteUserById(id);
+                if (user != null) {
+                    return new ResponseEntity<Void>(HttpStatus.OK);
+                } else { 
+                    return ResponseEntity.notFound().build();
+                }
+            } else {
+                return new ResponseEntity<Void>(HttpStatus.UNAUTHORIZED);
             }
         } 
         else 
@@ -88,21 +96,23 @@ public class UserApiController implements UserApi {
     public ResponseEntity<UserStatus> getStatsById(@ApiParam(value = "",required=true) @PathVariable("id") Long id,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "token", required = true) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            User user = userService.getUserById(id);
-            if (user != null) {
-                List<Borrowed> allBorrowed = libraryService.getAllBorrowedByUser(Long.valueOf(id));
-                List<Borrowed> damagedBorrowed = libraryService.getAllDamagedBorrowedBooksByList( allBorrowed );
-                List<Borrowed> currentBorrowed = libraryService.getAllCurrentBorrowedBooksByList( allBorrowed );
-                List<Borrowed> delayedBorrowed = libraryService.getAllDelayedBorrowedBooksByList( allBorrowed );
-                
-                UserStatus status = new UserStatus();
-                status.setNumberOfAllBorrowedBooks(allBorrowed.size());
-                status.setNumberOfAllDamagedBooks(damagedBorrowed.size());
-                status.setNumberOfCurrentBorrowedBooks(currentBorrowed.size());
-                status.setNumberOfDelayedBooks(delayedBorrowed.size());
-                return ResponseEntity.ok(status);
-            } else { 
-                return ResponseEntity.notFound().build();
+            if(userService.isViewPermittedForToken(id, token)){
+                User user = userService.getUserById(id);
+                if (user != null) {
+                    List<Borrowed> allBorrowed = libraryService.getAllBorrowedByUser(Long.valueOf(id));
+                    List<Borrowed> damagedBorrowed = libraryService.getAllDamagedBorrowedBooksByList( allBorrowed );
+                    List<Borrowed> currentBorrowed = libraryService.getAllCurrentBorrowedBooksByList( allBorrowed );
+                    List<Borrowed> delayedBorrowed = libraryService.getAllDelayedBorrowedBooksByList( allBorrowed );
+                    
+                    UserStatus status = new UserStatus();
+                    status.setNumberOfAllBorrowedBooks(allBorrowed.size());
+                    status.setNumberOfAllDamagedBooks(damagedBorrowed.size());
+                    status.setNumberOfCurrentBorrowedBooks(currentBorrowed.size());
+                    status.setNumberOfDelayedBooks(delayedBorrowed.size());
+                    return ResponseEntity.ok(status);
+                } else { 
+                    return ResponseEntity.notFound().build();
+                }
             }
         }
         return new ResponseEntity<UserStatus>(HttpStatus.BAD_REQUEST);
@@ -111,11 +121,13 @@ public class UserApiController implements UserApi {
     public ResponseEntity<User> getUserById(@ApiParam(value = "The id that needs to be fetched.",required=true) @PathVariable("id") Long id,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "token", required = true) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            User user = userService.getUserById(id);
-            if (user != null) {
-                return ResponseEntity.ok(user);
-            } else { 
-                return ResponseEntity.notFound().build();
+            if(userService.isViewPermittedForToken(id, token)){
+                User user = userService.getUserById(id);
+                if (user != null) {
+                    return ResponseEntity.ok(user);
+                } else { 
+                    return ResponseEntity.notFound().build();
+                }
             }
         }
         return new ResponseEntity<User>(HttpStatus.BAD_REQUEST);
@@ -127,7 +139,7 @@ public class UserApiController implements UserApi {
             User user = userService.loginUser(body.getEmail(), body.getPassword());
             if (user != null) {
                 HttpHeaders headers = new HttpHeaders();
-                headers.add("token", userTokenService.getTokenNameFormUserId(user.getId()));
+                headers.add("token", userService.getTokenNameFormUser(user));
                 return ResponseEntity.ok().headers(headers).body(user);
             } else { 
                 return ResponseEntity.notFound().build();
@@ -140,7 +152,7 @@ public class UserApiController implements UserApi {
     public ResponseEntity<Void> logoutUser(@ApiParam(value = "") @Valid @RequestParam(value = "token", required = false) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            Long userId = userTokenService.getUserIdFormToken(token);
+            Long userId = userService.getUserIdFormToken(token);
             if(userId != null){
                 userService.logoutUser(userId);
                 return new ResponseEntity<Void>(HttpStatus.OK);
@@ -154,11 +166,13 @@ public class UserApiController implements UserApi {
     public ResponseEntity<Void> updateUserById(@ApiParam(value = "Id that need to be updated",required=true) @PathVariable("id") Long id,@ApiParam(value = "Updated user object" ,required=true )  @Valid @RequestBody User body,@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "token", required = true) String token) {
         String accept = request.getHeader("Accept");
         if (accept != null && (accept.contains("application/json") || accept.contains("*/*")) ){
-            User user = userService.updateUserById(id, body);
-            if (user != null) {
-                return new ResponseEntity<Void>(HttpStatus.OK);
-            } else { 
-                return ResponseEntity.notFound().build();
+            if(userService.isModifyPermittedForToken(id, token)){
+                User user = userService.updateUserById(id, body);
+                if (user != null) {
+                    return new ResponseEntity<Void>(HttpStatus.OK);
+                } else { 
+                    return ResponseEntity.notFound().build();
+                }
             }
         }
         return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
